(bmcontainer) Create header file for bmcontainer
