+++ include/functions.h
@@ -1407,16 +1407,6 @@ struct Proc* StartTradeMenu(struct Unit* lUnit, struct Unit* rUnit, int unused);
 // ??? ActionPick(???);
 // ??? sub_803001C(???);
 // ??? sub_8030050(???);
-u16* GetConvoyItemArray(void);
-void ClearConvoyItems(void);
-void ShrinkConvoyItemList(void);
-int GetConvoyItemCount(void);
-int AddItemToConvoy(int);
-void RemoveItemFromConvoy(int);
-int GetConvoyItemSlot(int);
-bool8 HasConvoyAccess(void);
-bool8 sub_8031660(void);
-struct Unit* sub_8031688(void);
 // ??? InitUnitStack(???);
 // ??? PushUnit(???);
 // ??? LoadPlayerUnitsFromUnitStack(???);
